班级验证，学校验证bug修改

+++ api/src/main/java/com/mengyunzhi/springBootStudy/service/KlassServiceImpl.java
@@ -2,15 +2,22 @@ package com.mengyunzhi.springBootStudy.service;
 
 import com.mengyunzhi.springBootStudy.entity.Klass;
 import com.mengyunzhi.springBootStudy.entity.School;
+import com.mengyunzhi.springBootStudy.entity.Term;
+import com.mengyunzhi.springBootStudy.entity.User;
 import com.mengyunzhi.springBootStudy.repository.KlassRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 import javax.validation.constraints.NotNull;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * 班级服务实现
@@ -21,9 +28,32 @@ public class KlassServiceImpl implements KlassService {
     @Autowired
     KlassRepository klassRepository;
 
+    @Autowired
+    UserService userService;
+
+    @PersistenceContext
+    private EntityManager entityManager;
+
     @Override
-    public void deleteById(Long id) {
+    public ResponseEntity<Map<String, Object>> deleteById(Long id) {
+        Map<String, Object> response = new HashMap<>();
+
+        Klass klass = this.klassRepository.findById(id).get();
+
+        // 检查班级是否有用户
+        List<User> userList = this.userService.findByKlassId(klass.getId());
+        if (!userList.isEmpty()) {
+            response.put("status", "error");
+            response.put("message", "该班级仍有用户未清空");
+            return ResponseEntity.ok(response);
+        }
+
+        // 如果用户已经清空，删除班级
         this.klassRepository.deleteById(id);
+        response.put("status", "success");
+        response.put("message", "删除成功");
+        return ResponseEntity.ok(response);
+
     }
 
     @Override
@@ -63,8 +93,19 @@ public class KlassServiceImpl implements KlassService {
     }
 
     @Override
-    public void save(Klass klass) {
-        this.klassRepository.save(klass);
+    public ResponseEntity<Map<String, Object>> save(Klass klass) {
+        Map<String, Object> response = new HashMap<>();
+
+        if(!this.validateKlass(klass)){
+            response.put("status", "error");
+            response.put("message", "该班级已存在");
+            return ResponseEntity.ok(response);
+        } else {
+            this.klassRepository.save(klass);
+            response.put("status", "success");
+            response.put("message", "新增成功");
+            return ResponseEntity.ok(response);
+        }
     }
 
     /**
@@ -77,10 +118,35 @@ public class KlassServiceImpl implements KlassService {
      * @param klass 新班级数据
      */
     @Override
-    public void update(Long id, Klass klass) {
-        Klass oldKlass = klassRepository.findById(id).get();
-        oldKlass.setName(klass.getName());
-        oldKlass.setSchool(klass.getSchool());
-        klassRepository.save(oldKlass);
+    public ResponseEntity<Map<String, Object>> update(Long id, Klass klass) {
+        Klass oldKlass = this.klassRepository.findById(id).get();
+
+        Map<String, Object> response = new HashMap<>();
+
+        if(!this.validateKlass(klass)){
+            response.put("status", "error");
+            response.put("message", "该班级已存在");
+            return ResponseEntity.ok(response);
+        } else {
+            oldKlass.setName(klass.getName());
+            oldKlass.setSchool(klass.getSchool());
+            klassRepository.save(oldKlass);
+            response.put("status", "success");
+            response.put("message", "编辑成功");
+            return ResponseEntity.ok(response);
+        }
+    }
+
+    @Override
+    public boolean validateKlass(Klass klass) {
+        // 使用数据库查询，验证是否已经存在相同的Klass
+        String hql = "FROM Klass k WHERE k.name = :name AND k.school.id = :schoolId";
+        List<Klass> result = entityManager.createQuery(hql, Klass.class)
+                .setParameter("name", klass.getName())
+                .setParameter("schoolId", klass.getSchool().getId())
+                .getResultList();
+
+        // 如果查询结果不为空，则表示数据库中已经存在相同的Klass
+        return result.isEmpty();
     }
 }
